Create NOPAT
Return on Invested Capital is a measure of company efficiency. A given company is investing capital into itself to produce goods, those goods are then sold to generate profit. In general, ROIC is the amount of money a company makes that is above the cost of debt and equity. If ROIC is greater than WACC, it is efficient.

In general the game seems to be to think of a company as an investment machine. It generates profit and that profit should beat what the company could get investing its money elsewhere.